update form - fix input bug
