+++ src/app/storage/components/AddEditStorageModal/index.tsx
@@ -11,7 +11,7 @@ import {
   defaultAddEditStatus,
   createAddEditStatus,
   AddEditState,
- } from '../../../common/add_edit_state';
+} from '../../../common/add_edit_state';
 
 const AddEditStorageModal = ({
   addEditStatus,
@@ -27,7 +27,7 @@ const AddEditStorageModal = ({
     useState(initialStorageValues ? initialStorageValues.name : null);
 
   const onAddEditSubmit = (storageValues) => {
-    switch(addEditStatus.mode) {
+    switch (addEditStatus.mode) {
       case AddEditMode.Edit: {
         props.updateStorage(storageValues);
         break;
@@ -39,13 +39,13 @@ const AddEditStorageModal = ({
       }
       default: {
         console.warn(
-          `onAddEditSubmit, but unknown mode was found: ${addEditStatus.mode}. Ignoring.`, );
+          `onAddEditSubmit, but unknown mode was found: ${addEditStatus.mode}. Ignoring.`);
       }
     }
   };
 
   useEffect(() => {
-    if(isOpen && isPolling) {
+    if (isOpen && isPolling) {
       pollingContext.stopAllPolling();
     }
   });
@@ -67,7 +67,7 @@ const AddEditStorageModal = ({
   });
 
   return (
-    <Modal isSmall isOpen={isOpen} onClose={onClose} title={modalTitle}>
+    <Modal isSmall isOpen={isOpen} onClose={onClose} title={modalTitle} className="storage-modal-modifier">
       <AddEditStorageForm
         onAddEditSubmit={onAddEditSubmit}
         onClose={onClose}
@@ -96,7 +96,7 @@ export default connect(
     resetAddEditState: () => {
       dispatch(StorageActions.setStorageAddEditStatus(defaultAddEditStatus()));
     },
-    checkConnection: (storageName : string) => {
+    checkConnection: (storageName: string) => {
       dispatch(StorageActions.setStorageAddEditStatus(createAddEditStatus(
         AddEditState.Fetching, AddEditMode.Edit,
       )));
